Implement networking sessions base (#1)
Current implementation is basically `peer-to-peer` with the server
being a middleman (or data router, whatever you may call it).

Clients: Abstract `Serverside` class is somewhat of a plug-and-play
(not really). But close enough. A bit of lackluster implementation, not
sure how it will scale up.
Maybe renaming `Serverside` to `PeerShared` or something along that
line would describe it better for the time being.

Server: Multiple session management, each for two players is solid.

Object deletion is straight up not supported.